plant_disease_detection/requirements.txt
tensorflow==2.13.0
flask==2.3.3
gunicorn==20.1.0
numpy
scikit-learn
pandas
seaborn
matplotlib
Pillow
